Update script.js
I checked the blocking of downloading is wokring for me too.
And I suggest we can keep the async  on video clip for the cases when we don't lose the time to complete and download multiple clips.
In addition, I removed the overlapping variable, downloading. To do I changed the default value of conditnue_download from true to false.